FAC-375: Usecase renamed from LeaveSpaceMember to LeaveSpace
