+++ flickit-assessment-common/src/main/resources/i18n/users/messages_en.properties
@@ -87,4 +87,4 @@ create-space.title.notBlank='title' may not be empty
 create-space.title.size.min='title' length must be greater than or equal to 3
 create-space.title.size.max='title' length must be less than or equal to 100
 
-leave-space-member.spaceId.notNull='spaceId' may not be null
+leave-space.spaceId.notNull='spaceId' may not be null
