+++ flickit-assessment-users/src/main/java/org/flickit/assessment/users/common/ErrorMessageKey.java
@@ -81,5 +81,5 @@ public class ErrorMessageKey {
     public static final String CREATE_SPACE_TITLE_SIZE_MIN = "create-space.title.size.min";
     public static final String CREATE_SPACE_TITLE_SIZE_MAX = "create-space.title.size.max";
 
-    public  static final String LEAVE_SPACE_MEMBER_SPACE_ID_NOT_NULL = "leave-space-member.spaceId.notNull";
+    public  static final String LEAVE_SPACE_SPACE_ID_NOT_NULL = "leave-space.spaceId.notNull";
 }
