+++ flickit-assessment-users/src/test/java/org/flickit/assessment/users/application/port/in/spaceuseraccess/LeaveSpaceUseCaseParamTest.java
@@ -1,24 +1,24 @@
 package org.flickit.assessment.users.application.port.in.spaceuseraccess;
 
 import jakarta.validation.ConstraintViolationException;
-import org.flickit.assessment.users.application.port.in.spaceuseraccess.LeaveSpaceMemberUseCase.Param;
+import org.flickit.assessment.users.application.port.in.spaceuseraccess.LeaveSpaceUseCase.Param;
 import org.junit.jupiter.api.Test;
 
 import java.util.UUID;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.flickit.assessment.common.error.ErrorMessageKey.COMMON_CURRENT_USER_ID_NOT_NULL;
-import static org.flickit.assessment.users.common.ErrorMessageKey.LEAVE_SPACE_MEMBER_SPACE_ID_NOT_NULL;
+import static org.flickit.assessment.users.common.ErrorMessageKey.LEAVE_SPACE_SPACE_ID_NOT_NULL;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
-class LeaveSpaceMemberUseCaseParamTest {
+class LeaveSpaceUseCaseParamTest {
 
     @Test
     void testDeleteExpertGroupParam_idIsNull_ErrorMessage() {
         UUID currentUserId = UUID.randomUUID();
         var throwable = assertThrows(ConstraintViolationException.class,
             () -> new Param(null, currentUserId));
-        assertThat(throwable).hasMessage("id: " + LEAVE_SPACE_MEMBER_SPACE_ID_NOT_NULL);
+        assertThat(throwable).hasMessage("id: " + LEAVE_SPACE_SPACE_ID_NOT_NULL);
     }
 
     @Test
